Testing low-degree orientation in test_graph.py
